requirements.txt
pandas==2.2.3
websocket-client==1.8.0
pyinstaller==6.11.1
requests==2.32.3